Module stm32f4xx_hal::adc

Analog to digital converter configuration. According to CubeMx, all STM32F4 chips use the same ADC IP so this should be correct for all variants.

Modules

config
Contains types related to ADC configuration

Structs

Adc
Analog to Digital Converter

Temperature
Core temperature internal signal

Vbat
Vbat internal signal, used for monitoring the battery (if used)

Vref
Vref internal signal, used for calibration