Module stm32f4xx_hal::adc::config

Contains types related to ADC configuration

Structs

AdcConfig
Configuration for the adc. There are some additional parameters on the adc peripheral that can be added here when needed but this covers several basic usecases.

Enums

Align
Data register alignment

Clock
Clock config for the ADC Check the datasheet for the maximum speed the ADC supports

Continuous
Continuous mode enable/disable

Dma
DMA mode

Eoc
End-of-conversion interrupt enabled/disabled

ExternalTrigger
Possible external triggers the ADC can listen to

Resolution
Resolution to sample at

SampleTime
The number of cycles to sample a given channel for

Scan
Scan enable/disable

Sequence
The place in the sequence a given channel should be captured

TriggerMode
Possible trigger modes