Module stm32f4xx_hal::bb

Bit banding

Support for the manipulation of peripheral registers through bit-banding. Not all peripherals are mapped to the bit-banding alias region, the peripheral bit-band region is from 0x4000_0000 to 0x400F_FFFF. Bit-banding allows the manipulation of individual bits atomically.

Functions

clear⚠
Clears the bit on the provided register without modifying other bits.

set⚠
Sets the bit on the provided register without modifying other bits.

write⚠
Sets or clears the bit on the provided register without modifying other bits.