Trait stm32f4xx_hal::dma::traits::DMASet

pub unsafe trait DMASet<STREAM, CHANNEL, DIRECTION> { }

Trait to mark a set of Stream, Channel and Direction for a Peripheral as correct together.

Safety

Memory corruption might occur if this trait is implemented for an invalid combination.