Struct stm32f4xx_hal::dwt::Dwt

pub struct Dwt { /* fields omitted */ }

DWT (Data Watchpoint and Trace) unit

Implementations

impl Dwt

pub unsafe fn release(self) -> (DWT, DCB)

Release the dwt and dcb control

Safety

All instances of Delay and StopWatch become invalid after this

pub fn delay(&self) -> Delay

Create a delay instance

pub fn stopwatch<'i>(&self, times: &'i mut [u32]) -> StopWatch<'i>

Create a stopwatch instance

Arguments

times - Array which will be holding the timings in ticks (max laps == times.len()-1)

pub fn measure<F: FnOnce()>(&self, f: F) -> ClockDuration

Measure cycles it takes to execute closure f.

Since DWT Cycle Counter is a 32-bit counter that wraps around to 0 on overflow, users should be aware that Dwt::measure cannot correctly measure running time of closures which take longer than u32::MAX cycles