Struct stm32f4xx_hal::flash::UnlockedFlash

pub struct UnlockedFlash<'a> { /* fields omitted */ }

Result of FlashExt::unlocked()

Examples

use stm32f4xx_hal::pac::Peripherals;
use stm32f4xx_hal::flash::{FlashExt, LockedFlash, UnlockedFlash};
use embedded_storage::nor_flash::NorFlash;

let dp = Peripherals::take().unwrap();
let mut flash = LockedFlash::new(dp.FLASH);

// Unlock flash for writing
let mut unlocked_flash = flash.unlocked();

// Erase the second 128 KB sector.
NorFlash::erase(&mut unlocked_flash, 128 * 1024, 256 * 1024).unwrap();

// Write some data at the start of the second 128 KB sector.
let buf = [0u8; 64];
NorFlash::write(&mut unlocked_flash, 128 * 1024, &buf).unwrap();

// Lock flash by dropping
drop(unlocked_flash);

Implementations

impl UnlockedFlash<'_>

pub fn erase(&mut self, sector: u8) -> Result<(), Error>

Erase a flash sector

Refer to the reference manual to see which sector corresponds to which memory address.

pub fn program<'a, I>(&mut self, offset: usize, bytes: I) -> Result<(), Error> where
I: Iterator<Item = &'a u8>,

Program bytes with offset into flash memory, aligned to 128-bit rows