Struct stm32f4xx_hal::gpio::gpiob::PB4

pub struct PB4<MODE> { /* fields omitted */ }

Pin

Implementations

impl<MODE> PB4<MODE>

pub fn into_alternate_af0(self) -> PB4<Alternate<AF0>>

Configures the pin to operate in AF0 mode

pub fn into_alternate_af1(self) -> PB4<Alternate<AF1>>

Configures the pin to operate in AF1 mode

pub fn into_alternate_af2(self) -> PB4<Alternate<AF2>>

Configures the pin to operate in AF2 mode

pub fn into_alternate_af3(self) -> PB4<Alternate<AF3>>

Configures the pin to operate in AF3 mode

pub fn into_alternate_af4(self) -> PB4<Alternate<AF4>>

Configures the pin to operate in AF4 mode

pub fn into_alternate_af5(self) -> PB4<Alternate<AF5>>

Configures the pin to operate in AF5 mode

pub fn into_alternate_af6(self) -> PB4<Alternate<AF6>>

Configures the pin to operate in AF6 mode

pub fn into_alternate_af7(self) -> PB4<Alternate<AF7>>

Configures the pin to operate in AF7 mode

pub fn into_alternate_af8(self) -> PB4<Alternate<AF8>>

Configures the pin to operate in AF8 mode

pub fn into_alternate_af9(self) -> PB4<Alternate<AF9>>

Configures the pin to operate in AF9 mode

pub fn into_alternate_af10(self) -> PB4<Alternate<AF10>>

Configures the pin to operate in AF10 mode

pub fn into_alternate_af11(self) -> PB4<Alternate<AF11>>

Configures the pin to operate in AF11 mode

pub fn into_alternate_af12(self) -> PB4<Alternate<AF12>>

Configures the pin to operate in AF12 mode

pub fn into_alternate_af13(self) -> PB4<Alternate<AF13>>

Configures the pin to operate in AF13 mode

pub fn into_alternate_af14(self) -> PB4<Alternate<AF14>>

Configures the pin to operate in AF14 mode

pub fn into_alternate_af15(self) -> PB4<Alternate<AF15>>

Configures the pin to operate in AF15 mode

pub fn into_alternate_af0_open_drain(self) -> PB4<AlternateOD<AF0>>

Configures the pin to operate in AF0 open drain mode

pub fn into_alternate_af1_open_drain(self) -> PB4<AlternateOD<AF1>>

Configures the pin to operate in AF1 open drain mode

pub fn into_alternate_af2_open_drain(self) -> PB4<AlternateOD<AF2>>

Configures the pin to operate in AF2 open drain mode

pub fn into_alternate_af3_open_drain(self) -> PB4<AlternateOD<AF3>>

Configures the pin to operate in AF3 open drain mode

pub fn into_alternate_af4_open_drain(self) -> PB4<AlternateOD<AF4>>

Configures the pin to operate in AF4 open drain mode

pub fn into_alternate_af5_open_drain(self) -> PB4<AlternateOD<AF5>>

Configures the pin to operate in AF5 open drain mode

pub fn into_alternate_af6_open_drain(self) -> PB4<AlternateOD<AF6>>

Configures the pin to operate in AF6 open drain mode

pub fn into_alternate_af7_open_drain(self) -> PB4<AlternateOD<AF7>>

Configures the pin to operate in AF7 open drain mode

pub fn into_alternate_af8_open_drain(self) -> PB4<AlternateOD<AF8>>

Configures the pin to operate in AF8 open drain mode

pub fn into_alternate_af9_open_drain(self) -> PB4<AlternateOD<AF9>>

Configures the pin to operate in AF9 open drain mode

pub fn into_alternate_af10_open_drain(self) -> PB4<AlternateOD<AF10>>

Configures the pin to operate in AF10 open drain mode

pub fn into_alternate_af11_open_drain(self) -> PB4<AlternateOD<AF11>>

Configures the pin to operate in AF11 open drain mode

pub fn into_alternate_af12_open_drain(self) -> PB4<AlternateOD<AF12>>

Configures the pin to operate in AF12 open drain mode

pub fn into_alternate_af13_open_drain(self) -> PB4<AlternateOD<AF13>>

Configures the pin to operate in AF13 open drain mode

pub fn into_alternate_af14_open_drain(self) -> PB4<AlternateOD<AF14>>

Configures the pin to operate in AF14 open drain mode

pub fn into_alternate_af15_open_drain(self) -> PB4<AlternateOD<AF15>>

Configures the pin to operate in AF15 open drain mode

pub fn into_floating_input(self) -> PB4<Input<Floating>>

Configures the pin to operate as a floating input pin

pub fn into_pull_down_input(self) -> PB4<Input<PullDown>>

Configures the pin to operate as a pulled down input pin

pub fn into_pull_up_input(self) -> PB4<Input<PullUp>>

Configures the pin to operate as a pulled up input pin

pub fn into_open_drain_output(self) -> PB4<Output<OpenDrain>>

Configures the pin to operate as an open drain output pin

pub fn into_push_pull_output(self) -> PB4<Output<PushPull>>

Configures the pin to operate as an push pull output pin

pub fn into_analog(self) -> PB4<Analog>

Configures the pin to operate as an analog input pin

impl<MODE> PB4<Output<MODE>>

pub fn set_speed(self, speed: Speed) -> Self

Set pin speed

impl PB4<Output<OpenDrain>>

pub fn internal_pull_up(&mut self, on: bool)

Enables / disables the internal pull up

impl<MODE> PB4<Alternate<MODE>>

pub fn set_speed(self, speed: Speed) -> Self

Set pin speed

pub fn internal_pull_up(self, on: bool) -> Self

Enables / disables the internal pull up

impl<MODE> PB4<Alternate<MODE>>

pub fn set_open_drain(self) -> PB4<AlternateOD<MODE>>

Turns pin alternate configuration pin into open drain

impl<MODE> PB4<MODE>

pub fn downgrade(self) -> PB<MODE>

Erases the pin number from the type

This is useful when you want to collect the pins into an array where you need all the elements to have the same type