Struct stm32f4xx_hal::gpio::gpioe::PE6

pub struct PE6<MODE> { /* fields omitted */ }

Pin

Methods

impl<MODE> PE6<MODE>

pub fn into_alternate_af0(self) -> PE6<Alternate<AF0>>

Configures the pin to operate in AF0 mode

pub fn into_alternate_af1(self) -> PE6<Alternate<AF1>>

Configures the pin to operate in AF1 mode

pub fn into_alternate_af2(self) -> PE6<Alternate<AF2>>

Configures the pin to operate in AF2 mode

pub fn into_alternate_af3(self) -> PE6<Alternate<AF3>>

Configures the pin to operate in AF3 mode

pub fn into_alternate_af4(self) -> PE6<Alternate<AF4>>

Configures the pin to operate in AF4 mode

pub fn into_alternate_af5(self) -> PE6<Alternate<AF5>>

Configures the pin to operate in AF5 mode

pub fn into_alternate_af6(self) -> PE6<Alternate<AF6>>

Configures the pin to operate in AF6 mode

pub fn into_alternate_af7(self) -> PE6<Alternate<AF7>>

Configures the pin to operate in AF7 mode

pub fn into_alternate_af8(self) -> PE6<Alternate<AF8>>

Configures the pin to operate in AF8 mode

pub fn into_alternate_af9(self) -> PE6<Alternate<AF9>>

Configures the pin to operate in AF9 mode

pub fn into_alternate_af10(self) -> PE6<Alternate<AF10>>

Configures the pin to operate in AF10 mode

pub fn into_alternate_af11(self) -> PE6<Alternate<AF11>>

Configures the pin to operate in AF11 mode

pub fn into_alternate_af12(self) -> PE6<Alternate<AF12>>

Configures the pin to operate in AF12 mode

pub fn into_alternate_af13(self) -> PE6<Alternate<AF13>>

Configures the pin to operate in AF13 mode

pub fn into_alternate_af14(self) -> PE6<Alternate<AF14>>

Configures the pin to operate in AF14 mode

pub fn into_alternate_af15(self) -> PE6<Alternate<AF15>>

Configures the pin to operate in AF15 mode

pub fn into_floating_input(self) -> PE6<Input<Floating>>

Configures the pin to operate as a floating input pin

pub fn into_pull_down_input(self) -> PE6<Input<PullDown>>

Configures the pin to operate as a pulled down input pin

pub fn into_pull_up_input(self) -> PE6<Input<PullUp>>

Configures the pin to operate as a pulled up input pin

pub fn into_open_drain_output(self) -> PE6<Output<OpenDrain>>

Configures the pin to operate as an open drain output pin

pub fn into_push_pull_output(self) -> PE6<Output<PushPull>>

Configures the pin to operate as an push pull output pin

pub fn into_analog(self) -> PE6<Analog>

Configures the pin to operate as an analog input pin

impl<MODE> PE6<Output<MODE>>

pub fn set_speed(self, speed: Speed) -> Self

Set pin speed

impl PE6<Output<OpenDrain>>

pub fn internal_pull_up(&mut self, on: bool)

Enables / disables the internal pull up

impl<MODE> PE6<Alternate<MODE>>

pub fn set_speed(self, speed: Speed) -> Self

Set pin speed

pub fn internal_pull_up(self, on: bool) -> Self

Enables / disables the internal pull up

impl<MODE> PE6<Alternate<MODE>>

pub fn set_open_drain(self) -> Self

Turns pin alternate configuration pin into open drain

impl<MODE> PE6<MODE>

pub fn downgrade(self) -> PE<MODE>

Erases the pin number from the type

This is useful when you want to collect the pins into an array where you need all the elements to have the same type