Struct stm32f4xx_hal::gpio::Pin

pub struct Pin<MODE, const P: char, const N: u8> { /* fields omitted */ }

Generic pin type

MODE is one of the pin modes (see Modes section).

P is port name: A for GPIOA, B for GPIOB, etc.

N is pin number: from 0 to 15.

Implementations

impl<MODE, const P: char, const N: u8> Pin<MODE, P, N>

pub fn into_alternate<const A: u8>(self) -> Pin<Alternate<PushPull, A>, P, N>

Configures the pin to operate alternate mode

pub fn into_alternate_open_drain<const A: u8>(
self
) -> Pin<Alternate<OpenDrain, A>, P, N>

Configures the pin to operate in alternate open drain mode

pub fn into_floating_input(self) -> Pin<Input<Floating>, P, N>

Configures the pin to operate as a floating input pin

pub fn into_pull_down_input(self) -> Pin<Input<PullDown>, P, N>

Configures the pin to operate as a pulled down input pin

pub fn into_pull_up_input(self) -> Pin<Input<PullUp>, P, N>

Configures the pin to operate as a pulled up input pin

pub fn into_open_drain_output(self) -> Pin<Output<OpenDrain>, P, N>

Configures the pin to operate as an open drain output pin Initial state will be low.

pub fn into_open_drain_output_in_state(
self,
initial_state: PinState
) -> Pin<Output<OpenDrain>, P, N>

Configures the pin to operate as an open-drain output pin. initial_state specifies whether the pin should be initially high or low.

pub fn into_push_pull_output(self) -> Pin<Output<PushPull>, P, N>

Configures the pin to operate as an push pull output pin Initial state will be low.

pub fn into_push_pull_output_in_state(
self,
initial_state: PinState
) -> Pin<Output<PushPull>, P, N>

Configures the pin to operate as an push-pull output pin. initial_state specifies whether the pin should be initially high or low.

pub fn into_analog(self) -> Pin<Analog, P, N>

Configures the pin to operate as an analog input pin

pub fn into_dynamic(self) -> DynamicPin<P, N>

Configures the pin as a pin that can change between input and output without changing the type. It starts out as a floating input

impl<MODE, const P: char, const N: u8> Pin<MODE, P, N> where
MODE: PinMode,

pub fn with_floating_input<R>(
&mut self,
f: impl FnOnce(&mut Pin<Input<Floating>, P, N>) -> R
) -> R

Temporarily configures this pin as a floating input.

The closure f is called with the reconfigured pin. After it returns, the pin will be configured back.

pub fn with_pull_down_input<R>(
&mut self,
f: impl FnOnce(&mut Pin<Input<PullDown>, P, N>) -> R
) -> R

Temporarily configures this pin as a pulled-down input.

The closure f is called with the reconfigured pin. After it returns, the pin will be configured back.

pub fn with_pull_up_input<R>(
&mut self,
f: impl FnOnce(&mut Pin<Input<PullUp>, P, N>) -> R
) -> R

Temporarily configures this pin as a pulled-up input.

The closure f is called with the reconfigured pin. After it returns, the pin will be configured back.

pub fn with_analog<R>(
&mut self,
f: impl FnOnce(&mut Pin<Analog, P, N>) -> R
) -> R

Temporarily configures this pin as an analog pin.

The closure f is called with the reconfigured pin. After it returns, the pin will be configured back.

pub fn with_open_drain_output<R>(
&mut self,
f: impl FnOnce(&mut Pin<Output<OpenDrain>, P, N>) -> R
) -> R

Temporarily configures this pin as an open drain output.

The closure f is called with the reconfigured pin. After it returns, the pin will be configured back. The value of the pin after conversion is undefined. If you want to control it, use with_open_drain_output_in_state

pub fn with_open_drain_output_in_state<R>(
&mut self,
state: PinState,
f: impl FnOnce(&mut Pin<Output<OpenDrain>, P, N>) -> R
) -> R

Temporarily configures this pin as an open drain output .

The closure f is called with the reconfigured pin. After it returns, the pin will be configured back. Note that the new state is set slightly before conversion happens. This can cause a short output glitch if switching between output modes

pub fn with_push_pull_output<R>(
&mut self,
f: impl FnOnce(&mut Pin<Output<PushPull>, P, N>) -> R
) -> R

Temporarily configures this pin as a push-pull output.

The closure f is called with the reconfigured pin. After it returns, the pin will be configured back. The value of the pin after conversion is undefined. If you want to control it, use with_push_pull_output_in_state

pub fn with_push_pull_output_in_state<R>(
&mut self,
state: PinState,
f: impl FnOnce(&mut Pin<Output<PushPull>, P, N>) -> R
) -> R

Temporarily configures this pin as a push-pull output.

The closure f is called with the reconfigured pin. After it returns, the pin will be configured back. Note that the new state is set slightly before conversion happens. This can cause a short output glitch if switching between output modes

impl<MODE, const P: char, const N: u8> Pin<Output<MODE>, P, N>

pub fn set_speed(self, speed: Speed) -> Self

Set pin speed

impl<const P: char, const N: u8> Pin<Output<OpenDrain>, P, N>

pub fn internal_pull_up(self, on: bool) -> Self

Enables / disables the internal pull up

pub fn internal_pull_down(self, on: bool) -> Self

Enables / disables the internal pull down

impl<const P: char, const N: u8, const A: u8> Pin<Alternate<PushPull, A>, P, N>

pub fn set_speed(self, speed: Speed) -> Self

Set pin speed

pub fn internal_pull_up(self, on: bool) -> Self

Enables / disables the internal pull up

pub fn internal_pull_down(self, on: bool) -> Self

Enables / disables the internal pull down

impl<const P: char, const N: u8, const A: u8> Pin<Alternate<PushPull, A>, P, N>

pub fn set_open_drain(self) -> Pin<Alternate<OpenDrain, A>, P, N>

Turns pin alternate configuration pin into open drain

impl<MODE, const P: char, const N: u8> Pin<MODE, P, N>

pub fn erase_number(self) -> PEPin<MODE, P>

Erases the pin number from the type

This is useful when you want to collect the pins into an array where you need all the elements to have the same type

pub fn erase(self) -> EPin<MODE>

Erases the pin number and the port from the type

This is useful when you want to collect the pins into an array where you need all the elements to have the same type

impl<MODE, const P: char, const N: u8> Pin<Output<MODE>, P, N>

pub fn set_high(&mut self)

pub fn set_low(&mut self)

pub fn get_state(&self) -> PinState

pub fn set_state(&mut self, state: PinState)

pub fn is_set_high(&self) -> bool

pub fn is_set_low(&self) -> bool

pub fn toggle(&mut self)

impl<const P: char, const N: u8> Pin<Output<OpenDrain>, P, N>

pub fn is_high(&self) -> bool

pub fn is_low(&self) -> bool

impl<MODE, const P: char, const N: u8> Pin<Input<MODE>, P, N>

pub fn is_high(&self) -> bool

pub fn is_low(&self) -> bool