Module stm32f4xx_hal::pac::adc1::cr1

control register 1

Structs

AWDCH_R

Field AWDCH reader - Analog watchdog channel select bits

AWDCH_W

Field AWDCH writer - Analog watchdog channel select bits

AWDEN_R

Field AWDEN reader - Analog watchdog enable on regular channels

AWDEN_W

Field AWDEN writer - Analog watchdog enable on regular channels

AWDIE_R

Field AWDIE reader - Analog watchdog interrupt enable

AWDIE_W

Field AWDIE writer - Analog watchdog interrupt enable

AWDSGL_R

Field AWDSGL reader - Enable the watchdog on a single channel in scan mode

AWDSGL_W

Field AWDSGL writer - Enable the watchdog on a single channel in scan mode

CR1_SPEC

control register 1

DISCEN_R

Field DISCEN reader - Discontinuous mode on regular channels

DISCEN_W

Field DISCEN writer - Discontinuous mode on regular channels

DISCNUM_R

Field DISCNUM reader - Discontinuous mode channel count

DISCNUM_W

Field DISCNUM writer - Discontinuous mode channel count

EOCIE_R

Field EOCIE reader - Interrupt enable for EOC

EOCIE_W

Field EOCIE writer - Interrupt enable for EOC

JAUTO_R

Field JAUTO reader - Automatic injected group conversion

JAUTO_W

Field JAUTO writer - Automatic injected group conversion

JAWDEN_R

Field JAWDEN reader - Analog watchdog enable on injected channels

JAWDEN_W

Field JAWDEN writer - Analog watchdog enable on injected channels

JDISCEN_R

Field JDISCEN reader - Discontinuous mode on injected channels

JDISCEN_W

Field JDISCEN writer - Discontinuous mode on injected channels

JEOCIE_R

Field JEOCIE reader - Interrupt enable for injected channels

JEOCIE_W

Field JEOCIE writer - Interrupt enable for injected channels

OVRIE_R

Field OVRIE reader - Overrun interrupt enable

OVRIE_W

Field OVRIE writer - Overrun interrupt enable

R

Register CR1 reader

RES_R

Field RES reader - Resolution

RES_W

Field RES writer - Resolution

SCAN_R

Field SCAN reader - Scan mode

SCAN_W

Field SCAN writer - Scan mode

W

Register CR1 writer

Enums

AWDEN_A

Analog watchdog enable on regular channels

AWDIE_A

Analog watchdog interrupt enable

AWDSGL_A

Enable the watchdog on a single channel in scan mode

DISCEN_A

Discontinuous mode on regular channels

EOCIE_A

Interrupt enable for EOC

JAUTO_A

Automatic injected group conversion

JAWDEN_A

Analog watchdog enable on injected channels

JDISCEN_A

Discontinuous mode on injected channels

JEOCIE_A

Interrupt enable for injected channels

OVRIE_A

Overrun interrupt enable

RES_A

Resolution

SCAN_A

Scan mode