Struct stm32f4xx_hal::pac::dma1::hifcr::W

pub struct W(_);

Register HIFCR writer

Implementations

impl W

pub fn ctcif7(&mut self) -> CTCIF7_W<'_>

Bit 27 - Stream x clear transfer complete interrupt flag (x = 7..4)

pub fn chtif7(&mut self) -> CHTIF7_W<'_>

Bit 26 - Stream x clear half transfer interrupt flag (x = 7..4)

pub fn cteif7(&mut self) -> CTEIF7_W<'_>

Bit 25 - Stream x clear transfer error interrupt flag (x = 7..4)

pub fn cdmeif7(&mut self) -> CDMEIF7_W<'_>

Bit 24 - Stream x clear direct mode error interrupt flag (x = 7..4)

pub fn cfeif7(&mut self) -> CFEIF7_W<'_>

Bit 22 - Stream x clear FIFO error interrupt flag (x = 7..4)

pub fn ctcif6(&mut self) -> CTCIF6_W<'_>

Bit 21 - Stream x clear transfer complete interrupt flag (x = 7..4)

pub fn chtif6(&mut self) -> CHTIF6_W<'_>

Bit 20 - Stream x clear half transfer interrupt flag (x = 7..4)

pub fn cteif6(&mut self) -> CTEIF6_W<'_>

Bit 19 - Stream x clear transfer error interrupt flag (x = 7..4)

pub fn cdmeif6(&mut self) -> CDMEIF6_W<'_>

Bit 18 - Stream x clear direct mode error interrupt flag (x = 7..4)

pub fn cfeif6(&mut self) -> CFEIF6_W<'_>

Bit 16 - Stream x clear FIFO error interrupt flag (x = 7..4)

pub fn ctcif5(&mut self) -> CTCIF5_W<'_>

Bit 11 - Stream x clear transfer complete interrupt flag (x = 7..4)

pub fn chtif5(&mut self) -> CHTIF5_W<'_>

Bit 10 - Stream x clear half transfer interrupt flag (x = 7..4)

pub fn cteif5(&mut self) -> CTEIF5_W<'_>

Bit 9 - Stream x clear transfer error interrupt flag (x = 7..4)

pub fn cdmeif5(&mut self) -> CDMEIF5_W<'_>

Bit 8 - Stream x clear direct mode error interrupt flag (x = 7..4)

pub fn cfeif5(&mut self) -> CFEIF5_W<'_>

Bit 6 - Stream x clear FIFO error interrupt flag (x = 7..4)

pub fn ctcif4(&mut self) -> CTCIF4_W<'_>

Bit 5 - Stream x clear transfer complete interrupt flag (x = 7..4)

pub fn chtif4(&mut self) -> CHTIF4_W<'_>

Bit 4 - Stream x clear half transfer interrupt flag (x = 7..4)

pub fn cteif4(&mut self) -> CTEIF4_W<'_>

Bit 3 - Stream x clear transfer error interrupt flag (x = 7..4)

pub fn cdmeif4(&mut self) -> CDMEIF4_W<'_>

Bit 2 - Stream x clear direct mode error interrupt flag (x = 7..4)

pub fn cfeif4(&mut self) -> CFEIF4_W<'_>

Bit 0 - Stream x clear FIFO error interrupt flag (x = 7..4)

pub unsafe fn bits(&mut self, bits: u32) -> &mut W

Writes raw bits to the register.