Module stm32f4xx_hal::pac::ethernet_mmc

Ethernet: MAC management counters

Modules

mmccr
Ethernet MMC control register

mmcrfaecr
Ethernet MMC received frames with alignment error counter register

mmcrfcecr
Ethernet MMC received frames with CRC error counter register

mmcrgufcr
MMC received good unicast frames counter register

mmcrimr
Ethernet MMC receive interrupt mask register

mmcrir
Ethernet MMC receive interrupt register

mmctgfcr
Ethernet MMC transmitted good frames counter register

mmctgfmsccr
Ethernet MMC transmitted good frames after more than a single collision

mmctgfsccr
Ethernet MMC transmitted good frames after a single collision counter

mmctimr
Ethernet MMC transmit interrupt mask register

mmctir
Ethernet MMC transmit interrupt register

Structs

RegisterBlock
Register block

Type Definitions

MMCCR
Ethernet MMC control register

MMCRFAECR
Ethernet MMC received frames with alignment error counter register

MMCRFCECR
Ethernet MMC received frames with CRC error counter register

MMCRGUFCR
MMC received good unicast frames counter register

MMCRIMR
Ethernet MMC receive interrupt mask register

MMCRIR
Ethernet MMC receive interrupt register

MMCTGFCR
Ethernet MMC transmitted good frames counter register

MMCTGFMSCCR
Ethernet MMC transmitted good frames after more than a single collision

MMCTGFSCCR
Ethernet MMC transmitted good frames after a single collision counter

MMCTIMR
Ethernet MMC transmit interrupt mask register

MMCTIR
Ethernet MMC transmit interrupt register