Module stm32f4xx_hal::pac::gpioa::afrh

GPIO alternate function high register

Structs

AFRH8_W
Write proxy for field AFRH8

AFRH9_W
Write proxy for field AFRH9

AFRH10_W
Write proxy for field AFRH10

AFRH11_W
Write proxy for field AFRH11

AFRH12_W
Write proxy for field AFRH12

AFRH13_W
Write proxy for field AFRH13

AFRH14_W
Write proxy for field AFRH14

AFRH15_W
Write proxy for field AFRH15

Enums

AFRH15_A
Alternate function selection for port x bit y (y = 8..15)

Type Definitions

AFRH8_A
Alternate function selection for port x bit y (y = 8..15)

AFRH8_R
Reader of field AFRH8

AFRH9_A
Alternate function selection for port x bit y (y = 8..15)

AFRH9_R
Reader of field AFRH9

AFRH10_A
Alternate function selection for port x bit y (y = 8..15)

AFRH10_R
Reader of field AFRH10

AFRH11_A
Alternate function selection for port x bit y (y = 8..15)

AFRH11_R
Reader of field AFRH11

AFRH12_A
Alternate function selection for port x bit y (y = 8..15)

AFRH12_R
Reader of field AFRH12

AFRH13_A
Alternate function selection for port x bit y (y = 8..15)

AFRH13_R
Reader of field AFRH13

AFRH14_A
Alternate function selection for port x bit y (y = 8..15)

AFRH14_R
Reader of field AFRH14

AFRH15_R
Reader of field AFRH15

R
Reader of register AFRH

W
Writer for register AFRH