Module stm32f4xx_hal::pac::sdio::sta

status register

Structs

CCRCFAIL_R

Field CCRCFAIL reader - Command response received (CRC check failed)

CEATAEND_R

Field CEATAEND reader - CE-ATA command completion signal received for CMD61

CMDACT_R

Field CMDACT reader - Command transfer in progress

CMDREND_R

Field CMDREND reader - Command response received (CRC check passed)

CMDSENT_R

Field CMDSENT reader - Command sent (no response required)

CTIMEOUT_R

Field CTIMEOUT reader - Command response timeout

DATAEND_R

Field DATAEND reader - Data end (data counter, SDIDCOUNT, is zero)

DBCKEND_R

Field DBCKEND reader - Data block sent/received (CRC check passed)

DCRCFAIL_R

Field DCRCFAIL reader - Data block sent/received (CRC check failed)

DTIMEOUT_R

Field DTIMEOUT reader - Data timeout

R

Register STA reader

RXACT_R

Field RXACT reader - Data receive in progress

RXDAVL_R

Field RXDAVL reader - Data available in receive FIFO

RXFIFOE_R

Field RXFIFOE reader - Receive FIFO empty

RXFIFOF_R

Field RXFIFOF reader - Receive FIFO full

RXFIFOHF_R

Field RXFIFOHF reader - Receive FIFO half full: there are at least 8 words in the FIFO

RXOVERR_R

Field RXOVERR reader - Received FIFO overrun error

SDIOIT_R

Field SDIOIT reader - SDIO interrupt received

STA_SPEC

status register

STBITERR_R

Field STBITERR reader - Start bit not detected on all data signals in wide bus mode

TXACT_R

Field TXACT reader - Data transmit in progress

TXDAVL_R

Field TXDAVL reader - Data available in transmit FIFO

TXFIFOE_R

Field TXFIFOE reader - Transmit FIFO empty

TXFIFOF_R

Field TXFIFOF reader - Transmit FIFO full

TXFIFOHE_R

Field TXFIFOHE reader - Transmit FIFO half empty: at least 8 words can be written into the FIFO

TXUNDERR_R

Field TXUNDERR reader - Transmit FIFO underrun error

Enums

CCRCFAIL_A

Command response received (CRC check failed)

CEATAEND_A

CE-ATA command completion signal received for CMD61

CMDACT_A

Command transfer in progress

CMDREND_A

Command response received (CRC check passed)

CMDSENT_A

Command sent (no response required)

CTIMEOUT_A

Command response timeout

DATAEND_A

Data end (data counter, SDIDCOUNT, is zero)

DBCKEND_A

Data block sent/received (CRC check passed)

DCRCFAIL_A

Data block sent/received (CRC check failed)

DTIMEOUT_A

Data timeout

RXACT_A

Data receive in progress

RXDAVL_A

Data available in receive FIFO

RXFIFOE_A

Receive FIFO empty

RXFIFOF_A

Receive FIFO full

RXFIFOHF_A

Receive FIFO half full: there are at least 8 words in the FIFO

RXOVERR_A

Received FIFO overrun error

SDIOIT_A

SDIO interrupt received

STBITERR_A

Start bit not detected on all data signals in wide bus mode

TXACT_A

Data transmit in progress

TXDAVL_A

Data available in transmit FIFO

TXFIFOE_A

Transmit FIFO empty

TXFIFOF_A

Transmit FIFO full

TXFIFOHE_A

Transmit FIFO half empty: at least 8 words can be written into the FIFO

TXUNDERR_A

Transmit FIFO underrun error