Trait stm32f4xx_hal::prelude::_embedded_hal_blocking_delay_DelayMs

pub trait _embedded_hal_blocking_delay_DelayMs<UXX> {
    fn delay_ms(&mut self, ms: UXX);
}

Millisecond delay

UXX denotes the range type of the delay time. UXX can be u8, u16, etc. A single type can implement this trait for different types of UXX.

Required methods

fn delay_ms(&mut self, ms: UXX)

Pauses execution for ms milliseconds