Trait stm32f4xx_hal::prelude::_embedded_hal_digital_v2_OutputPin

pub trait _embedded_hal_digital_v2_OutputPin {
    type Error;
    fn set_low(&mut self) -> Result<(), Self::Error>;

    fn set_high(&mut self) -> Result<(), Self::Error>;
}

Single digital push-pull output pin

Associated Types

type Error

Error type

Loading content...
Required methods

fn set_low(&mut self) -> Result<(), Self::Error>

Drives the pin low

NOTE the actual electrical state of the pin may not actually be low, e.g. due to external electrical sources

fn set_high(&mut self) -> Result<(), Self::Error>

Drives the pin high

NOTE the actual electrical state of the pin may not actually be high, e.g. due to external electrical sources

Loading content...