Struct stm32f4xx_hal::rcc::CFGR

pub struct CFGR { /* fields omitted */ }

Implementations

impl CFGR

pub fn use_hse(self, freq: Hertz) -> Self

Uses HSE (external oscillator) instead of HSI (internal RC oscillator) as the clock source. Will result in a hang if an external oscillator is not connected or it fails to start.

pub fn bypass_hse_oscillator(self) -> Self

Bypasses the high-speed external oscillator and uses an external clock input on the OSC_IN pin.

For this configuration, the OSC_IN pin should be connected to a clock source with a frequency specified in the call to use_hse(), and the OSC_OUT pin should not be connected.

This function has no effect unless use_hse() is also called.

pub fn hclk(self, freq: Hertz) -> Self

pub fn pclk1(self, freq: Hertz) -> Self

pub fn pclk2(self, freq: Hertz) -> Self

pub fn sysclk(self, freq: Hertz) -> Self

pub fn require_pll48clk(self) -> Self

pub fn i2s_ckin(self, freq: Hertz) -> Self

Declares that the selected frequency is available at the I2S clock input pin (I2S_CKIN).

If this frequency matches the requested SAI or I2S frequencies, the external I2S clock is used to generate the clocks.

pub fn i2s_clk(self, freq: Hertz) -> Self

Selects an I2S clock frequency and enables the I2S clock.

pub fn freeze(self) -> Clocks

Initialises the hardware according to CFGR state returning a Clocks instance. Panics if overclocking is attempted.

pub unsafe fn freeze_unchecked(self) -> Clocks

Initialises the hardware according to CFGR state returning a Clocks instance. Allows overclocking.

Safety

This method does not check if the clocks are bigger or smaller than the officially recommended.