Module stm32f4xx_hal::serial

Modules

config

Structs

NoRx
A filler type for when the Rx pin is unnecessary

NoTx
A filler type for when the Tx pin is unnecessary

Rx
Serial receiver

Serial
Serial abstraction

Tx
Serial transmitter

Enums

Error
Serial error

Event
Interrupt event

Traits

PinRx

PinTx

Pins