Struct stm32f4xx_hal::serial::Rx

pub struct Rx<USART, WORD = u8> { /* fields omitted */ }

Serial receiver

Implementations

impl<USART, WORD> Rx<USART, WORD> where
USART: Instance,

pub fn listen(&mut self)

Start listening for an rx not empty interrupt event

Note, you will also have to enable the corresponding interrupt in the NVIC to start receiving events.

pub fn unlisten(&mut self)

Stop listening for the rx not empty interrupt event

pub fn listen_idle(&mut self)

Start listening for a line idle interrupt event

Note, you will also have to enable the corresponding interrupt in the NVIC to start receiving events.

pub fn unlisten_idle(&mut self)

Stop listening for the line idle interrupt event

pub fn is_idle(&self) -> bool

Return true if the line idle status is set

pub fn is_rx_not_empty(&self) -> bool

Return true if the rx register is not empty (and can be read)

pub fn clear_idle_interrupt(&self)

Clear idle line interrupt flag