Module stm32f4xx_hal::signature

Device electronic signature

(stored in flash memory)

Structs

FlashSize
Size of integrated flash

Uid
Uniqure Device ID register

VrefCal
ADC VREF calibration value is stored in at the factory

VtempCal30
A temperature reading taken at 30°C stored at the factory

VtempCal110
A temperature reading taken at 110°C stored at the factory

Constants

VDDA_CALIB
This is the test voltage, in millivolts of the calibration done at the factory