Module stm32f4xx_hal::spi

Structs

Mode
SPI mode

NoMiso
A filler type for when the Miso pin is unnecessary

NoMosi
A filler type for when the Mosi pin is unnecessary

NoSck
A filler type for when the SCK pin is unnecessary

Spi

Enums

Error
SPI error

Event
Interrupt events

Phase
Clock phase

Polarity
Clock polarity

Traits

PinMiso

PinMosi

PinSck

Pins