Module stm32f4xx_hal::stm32::gpioa::otyper

GPIO port output type register

Structs

R
Value read from the register

W
Value to write to the register

_OT0W
Proxy

_OT1W
Proxy

_OT2W
Proxy

_OT3W
Proxy

_OT4W
Proxy

_OT5W
Proxy

_OT6W
Proxy

_OT7W
Proxy

_OT8W
Proxy

_OT9W
Proxy

_OT10W
Proxy

_OT11W
Proxy

_OT12W
Proxy

_OT13W
Proxy

_OT14W
Proxy

_OT15W
Proxy

Enums

OT15R
Possible values of the field OT15

OT15W
Values that can be written to the field OT15

Type Definitions

OT0R
Possible values of the field OT0

OT0W
Values that can be written to the field OT0

OT1R
Possible values of the field OT1

OT1W
Values that can be written to the field OT1

OT2R
Possible values of the field OT2

OT2W
Values that can be written to the field OT2

OT3R
Possible values of the field OT3

OT3W
Values that can be written to the field OT3

OT4R
Possible values of the field OT4

OT4W
Values that can be written to the field OT4

OT5R
Possible values of the field OT5

OT5W
Values that can be written to the field OT5

OT6R
Possible values of the field OT6

OT6W
Values that can be written to the field OT6

OT7R
Possible values of the field OT7

OT7W
Values that can be written to the field OT7

OT8R
Possible values of the field OT8

OT8W
Values that can be written to the field OT8

OT9R
Possible values of the field OT9

OT9W
Values that can be written to the field OT9

OT10R
Possible values of the field OT10

OT10W
Values that can be written to the field OT10

OT11R
Possible values of the field OT11

OT11W
Values that can be written to the field OT11

OT12R
Possible values of the field OT12

OT12W
Values that can be written to the field OT12

OT13R
Possible values of the field OT13

OT13W
Values that can be written to the field OT13

OT14R
Possible values of the field OT14

OT14W
Values that can be written to the field OT14