Module stm32f4xx_hal::stm32::i2c1::cr1

Control register 1

Structs

R
Value read from the register

W
Value to write to the register

_ACKW
Proxy

_ALERTW
Proxy

_ENARPW
Proxy

_ENGCW
Proxy

_ENPECW
Proxy

_NOSTRETCHW
Proxy

_PECW
Proxy

_PEW
Proxy

_POSW
Proxy

_SMBTYPEW
Proxy

_SMBUSW
Proxy

_STARTW
Proxy

_STOPW
Proxy

_SWRSTW
Proxy

Enums

ACKR
Possible values of the field ACK

ACKW
Values that can be written to the field ACK

ALERTR
Possible values of the field ALERT

ALERTW
Values that can be written to the field ALERT

ENARPR
Possible values of the field ENARP

ENARPW
Values that can be written to the field ENARP

ENGCR
Possible values of the field ENGC

ENGCW
Values that can be written to the field ENGC

ENPECR
Possible values of the field ENPEC

ENPECW
Values that can be written to the field ENPEC

NOSTRETCHR
Possible values of the field NOSTRETCH

NOSTRETCHW
Values that can be written to the field NOSTRETCH

PECR
Possible values of the field PEC

PECW
Values that can be written to the field PEC

PER
Possible values of the field PE

PEW
Values that can be written to the field PE

POSR
Possible values of the field POS

POSW
Values that can be written to the field POS

SMBTYPER
Possible values of the field SMBTYPE

SMBTYPEW
Values that can be written to the field SMBTYPE

SMBUSR
Possible values of the field SMBUS

SMBUSW
Values that can be written to the field SMBUS

STARTR
Possible values of the field START

STARTW
Values that can be written to the field START

STOPR
Possible values of the field STOP

STOPW
Values that can be written to the field STOP

SWRSTR
Possible values of the field SWRST

SWRSTW
Values that can be written to the field SWRST