Module stm32f4xx_hal::stm32::rcc::apb1lpenr

APB1 peripheral clock enable in low power mode register

Structs

R
Value read from the register

W
Value to write to the register

_I2C1LPENW
Proxy

_I2C2LPENW
Proxy

_I2C3LPENW
Proxy

_PWRLPENW
Proxy

_SPI2LPENW
Proxy

_SPI3LPENW
Proxy

_TIM2LPENW
Proxy

_TIM3LPENW
Proxy

_TIM4LPENW
Proxy

_TIM5LPENW
Proxy

_USART2LPENW
Proxy

_WWDGLPENW
Proxy

Enums

PWRLPENR
Possible values of the field PWRLPEN

PWRLPENW
Values that can be written to the field PWRLPEN

Type Definitions

I2C1LPENR
Possible values of the field I2C1LPEN

I2C1LPENW
Values that can be written to the field I2C1LPEN

I2C2LPENR
Possible values of the field I2C2LPEN

I2C2LPENW
Values that can be written to the field I2C2LPEN

I2C3LPENR
Possible values of the field I2C3LPEN

I2C3LPENW
Values that can be written to the field I2C3LPEN

SPI2LPENR
Possible values of the field SPI2LPEN

SPI2LPENW
Values that can be written to the field SPI2LPEN

SPI3LPENR
Possible values of the field SPI3LPEN

SPI3LPENW
Values that can be written to the field SPI3LPEN

TIM2LPENR
Possible values of the field TIM2LPEN

TIM2LPENW
Values that can be written to the field TIM2LPEN

TIM3LPENR
Possible values of the field TIM3LPEN

TIM3LPENW
Values that can be written to the field TIM3LPEN

TIM4LPENR
Possible values of the field TIM4LPEN

TIM4LPENW
Values that can be written to the field TIM4LPEN

TIM5LPENR
Possible values of the field TIM5LPEN

TIM5LPENW
Values that can be written to the field TIM5LPEN

USART2LPENR
Possible values of the field USART2LPEN

USART2LPENW
Values that can be written to the field USART2LPEN

WWDGLPENR
Possible values of the field WWDGLPEN

WWDGLPENW
Values that can be written to the field WWDGLPEN