Module stm32f4xx_hal::stm32::rcc::apb2lpenr

APB2 peripheral clock enabled in low power mode register

Structs

EXTITENR
Value of the field

R
Value read from the register

W
Value to write to the register

_ADC1LPENW
Proxy

_DFSDM2LPENW
Proxy

_DFSDMLPENW
Proxy

_EXTITENW
Proxy

_SAI1LPENW
Proxy

_SDIOLPENW
Proxy

_SPI1LPENW
Proxy

_SPI4LPENW
Proxy

_SPI5LPENW
Proxy

_SYSCFGLPENW
Proxy

_TIM1LPENW
Proxy

_TIM8LPENW
Proxy

_TIM9LPENW
Proxy

_TIM10LPENW
Proxy

_TIM11LPENW
Proxy

_UART9LPENW
Proxy

_UART10LPENW
Proxy

_USART1LPENW
Proxy

_USART6LPENW
Proxy

_USART9LPENW
Proxy

_USART10LPENW
Proxy

Enums

TIM1LPENR
Possible values of the field TIM1LPEN

TIM1LPENW
Values that can be written to the field TIM1LPEN

Type Definitions

ADC1LPENR
Possible values of the field ADC1LPEN

ADC1LPENW
Values that can be written to the field ADC1LPEN

DFSDM2LPENR
Possible values of the field DFSDM2LPEN

DFSDM2LPENW
Values that can be written to the field DFSDM2LPEN

DFSDMLPENR
Possible values of the field DFSDMLPEN

DFSDMLPENW
Values that can be written to the field DFSDMLPEN

SAI1LPENR
Possible values of the field SAI1LPEN

SAI1LPENW
Values that can be written to the field SAI1LPEN

SDIOLPENR
Possible values of the field SDIOLPEN

SDIOLPENW
Values that can be written to the field SDIOLPEN

SPI1LPENR
Possible values of the field SPI1LPEN

SPI1LPENW
Values that can be written to the field SPI1LPEN

SPI4LPENR
Possible values of the field SPI4LPEN

SPI4LPENW
Values that can be written to the field SPI4LPEN

SPI5LPENR
Possible values of the field SPI5LPEN

SPI5LPENW
Values that can be written to the field SPI5LPEN

SYSCFGLPENR
Possible values of the field SYSCFGLPEN

SYSCFGLPENW
Values that can be written to the field SYSCFGLPEN

TIM8LPENR
Possible values of the field TIM8LPEN

TIM8LPENW
Values that can be written to the field TIM8LPEN

TIM9LPENR
Possible values of the field TIM9LPEN

TIM9LPENW
Values that can be written to the field TIM9LPEN

TIM10LPENR
Possible values of the field TIM10LPEN

TIM10LPENW
Values that can be written to the field TIM10LPEN

TIM11LPENR
Possible values of the field TIM11LPEN

TIM11LPENW
Values that can be written to the field TIM11LPEN

UART9LPENR
Possible values of the field UART9LPEN

UART9LPENW
Values that can be written to the field UART9LPEN

UART10LPENR
Possible values of the field UART10LPEN

UART10LPENW
Values that can be written to the field UART10LPEN

USART1LPENR
Possible values of the field USART1LPEN

USART1LPENW
Values that can be written to the field USART1LPEN

USART6LPENR
Possible values of the field USART6LPEN

USART6LPENW
Values that can be written to the field USART6LPEN

USART9LPENR
Possible values of the field USART9LPEN

USART9LPENW
Values that can be written to the field USART9LPEN

USART10LPENR
Possible values of the field USART10LPEN

USART10LPENW
Values that can be written to the field USART10LPEN