Module stm32f4xx_hal::stm32::rcc::cfgr

clock configuration register

Structs

R
Value read from the register

RTCPRER
Value of the field

W
Value to write to the register

_HPREW
Proxy

_MCO1PREW
Proxy

_MCO1W
Proxy

_MCO2W
Proxy

_MCO2PREW
Proxy

_PPRE1W
Proxy

_PPRE2W
Proxy

_RTCPREW
Proxy

_SWSW
Proxy

_SWW
Proxy

Enums

HPRER
Possible values of the field HPRE

HPREW
Values that can be written to the field HPRE

MCO1R
Possible values of the field MCO1

MCO1W
Values that can be written to the field MCO1

MCO2R
Possible values of the field MCO2

MCO2PRER
Possible values of the field MCO2PRE

MCO2W
Values that can be written to the field MCO2

MCO2PREW
Values that can be written to the field MCO2PRE

PPRE2R
Possible values of the field PPRE2

PPRE2W
Values that can be written to the field PPRE2

SWR
Possible values of the field SW

SWSR
Possible values of the field SWS

SWSW
Values that can be written to the field SWS

SWW
Values that can be written to the field SW

Type Definitions

MCO1PRER
Possible values of the field MCO1PRE

MCO1PREW
Values that can be written to the field MCO1PRE

PPRE1R
Possible values of the field PPRE1

PPRE1W
Values that can be written to the field PPRE1