Enum stm32f4xx_hal::stm32::rcc::dckcfgr::PLLDIVRR

pub enum PLLDIVRR {
    DIV1,
    DIV2,
    DIV3,
    DIV4,
    DIV5,
    DIV6,
    DIV7,
    DIV8,
    DIV9,
    DIV10,
    DIV11,
    DIV12,
    DIV13,
    DIV14,
    DIV15,
    DIV16,
    DIV17,
    DIV18,
    DIV19,
    DIV20,
    DIV21,
    DIV22,
    DIV23,
    DIV24,
    DIV25,
    DIV26,
    DIV27,
    DIV28,
    DIV29,
    DIV30,
    DIV31,
    DIV32,
}

Possible values of the field PLLDIVR

Variants
DIV1

PLLSAIDIVQ = /1

DIV2

PLLSAIDIVQ = /2

DIV3

PLLSAIDIVQ = /3

DIV4

PLLSAIDIVQ = /4

DIV5

PLLSAIDIVQ = /5

DIV6

PLLSAIDIVQ = /6

DIV7

PLLSAIDIVQ = /7

DIV8

PLLSAIDIVQ = /8

DIV9

PLLSAIDIVQ = /9

DIV10

PLLSAIDIVQ = /10

DIV11

PLLSAIDIVQ = /11

DIV12

PLLSAIDIVQ = /12

DIV13

PLLSAIDIVQ = /13

DIV14

PLLSAIDIVQ = /14

DIV15

PLLSAIDIVQ = /15

DIV16

PLLSAIDIVQ = /16

DIV17

PLLSAIDIVQ = /17

DIV18

PLLSAIDIVQ = /18

DIV19

PLLSAIDIVQ = /19

DIV20

PLLSAIDIVQ = /20

DIV21

PLLSAIDIVQ = /21

DIV22

PLLSAIDIVQ = /22

DIV23

PLLSAIDIVQ = /23

DIV24

PLLSAIDIVQ = /24

DIV25

PLLSAIDIVQ = /25

DIV26

PLLSAIDIVQ = /26

DIV27

PLLSAIDIVQ = /27

DIV28

PLLSAIDIVQ = /28

DIV29

PLLSAIDIVQ = /29

DIV30

PLLSAIDIVQ = /30

DIV31

PLLSAIDIVQ = /31

DIV32

PLLSAIDIVQ = /32

Methods

impl PLLDIVRR

pub fn bits(&self) -> u8

Value of the field as raw bits

pub fn is_div1(&self) -> bool

Checks if the value of the field is DIV1

pub fn is_div2(&self) -> bool

Checks if the value of the field is DIV2

pub fn is_div3(&self) -> bool

Checks if the value of the field is DIV3

pub fn is_div4(&self) -> bool

Checks if the value of the field is DIV4

pub fn is_div5(&self) -> bool

Checks if the value of the field is DIV5

pub fn is_div6(&self) -> bool

Checks if the value of the field is DIV6

pub fn is_div7(&self) -> bool

Checks if the value of the field is DIV7

pub fn is_div8(&self) -> bool

Checks if the value of the field is DIV8

pub fn is_div9(&self) -> bool

Checks if the value of the field is DIV9

pub fn is_div10(&self) -> bool

Checks if the value of the field is DIV10

pub fn is_div11(&self) -> bool

Checks if the value of the field is DIV11

pub fn is_div12(&self) -> bool

Checks if the value of the field is DIV12

pub fn is_div13(&self) -> bool

Checks if the value of the field is DIV13

pub fn is_div14(&self) -> bool

Checks if the value of the field is DIV14

pub fn is_div15(&self) -> bool

Checks if the value of the field is DIV15

pub fn is_div16(&self) -> bool

Checks if the value of the field is DIV16

pub fn is_div17(&self) -> bool

Checks if the value of the field is DIV17

pub fn is_div18(&self) -> bool

Checks if the value of the field is DIV18

pub fn is_div19(&self) -> bool

Checks if the value of the field is DIV19

pub fn is_div20(&self) -> bool

Checks if the value of the field is DIV20

pub fn is_div21(&self) -> bool

Checks if the value of the field is DIV21

pub fn is_div22(&self) -> bool

Checks if the value of the field is DIV22

pub fn is_div23(&self) -> bool

Checks if the value of the field is DIV23

pub fn is_div24(&self) -> bool

Checks if the value of the field is DIV24

pub fn is_div25(&self) -> bool

Checks if the value of the field is DIV25

pub fn is_div26(&self) -> bool

Checks if the value of the field is DIV26

pub fn is_div27(&self) -> bool

Checks if the value of the field is DIV27

pub fn is_div28(&self) -> bool

Checks if the value of the field is DIV28

pub fn is_div29(&self) -> bool

Checks if the value of the field is DIV29

pub fn is_div30(&self) -> bool

Checks if the value of the field is DIV30

pub fn is_div31(&self) -> bool

Checks if the value of the field is DIV31

pub fn is_div32(&self) -> bool

Checks if the value of the field is DIV32