Module stm32f4xx_hal::stm32::spi5::cr2

control register 2

Structs

R
Value read from the register

W
Value to write to the register

_ERRIEW
Proxy

_FRFW
Proxy

_RXDMAENW
Proxy

_RXNEIEW
Proxy

_SSOEW
Proxy

_TXDMAENW
Proxy

_TXEIEW
Proxy

Enums

ERRIER
Possible values of the field ERRIE

ERRIEW
Values that can be written to the field ERRIE

FRFR
Possible values of the field FRF

FRFW
Values that can be written to the field FRF

RXDMAENR
Possible values of the field RXDMAEN

RXDMAENW
Values that can be written to the field RXDMAEN

RXNEIER
Possible values of the field RXNEIE

RXNEIEW
Values that can be written to the field RXNEIE

SSOER
Possible values of the field SSOE

SSOEW
Values that can be written to the field SSOE

TXDMAENR
Possible values of the field TXDMAEN

TXDMAENW
Values that can be written to the field TXDMAEN

TXEIER
Possible values of the field TXEIE

TXEIEW
Values that can be written to the field TXEIE