Module stm32f4xx_hal::stm32::tim1::cr1

control register 1

Structs

CKDR
Value of the field

R
Value read from the register

W
Value to write to the register

_ARPEW
Proxy

_CENW
Proxy

_CKDW
Proxy

_CMSW
Proxy

_DIRW
Proxy

_OPMW
Proxy

_UDISW
Proxy

_URSW
Proxy

Enums

ARPER
Possible values of the field ARPE

ARPEW
Values that can be written to the field ARPE

CENR
Possible values of the field CEN

CENW
Values that can be written to the field CEN

CMSR
Possible values of the field CMS

CMSW
Values that can be written to the field CMS

DIRR
Possible values of the field DIR

DIRW
Values that can be written to the field DIR

OPMR
Possible values of the field OPM

OPMW
Values that can be written to the field OPM

UDISR
Possible values of the field UDIS

UDISW
Values that can be written to the field UDIS

URSR
Possible values of the field URS

URSW
Values that can be written to the field URS