Module stm32f4xx_hal::stm32::usart3::cr2

Control register 2

Structs

ADDR
Value of the field

LBCLR
Value of the field

R
Value read from the register

W
Value to write to the register

_ADDW
Proxy

_CLKENW
Proxy

_CPHAW
Proxy

_CPOLW
Proxy

_LBCLW
Proxy

_LBDIEW
Proxy

_LBDLW
Proxy

_LINENW
Proxy

_STOPW
Proxy

Enums

CLKENR
Possible values of the field CLKEN

CLKENW
Values that can be written to the field CLKEN

CPHAR
Possible values of the field CPHA

CPHAW
Values that can be written to the field CPHA

CPOLR
Possible values of the field CPOL

CPOLW
Values that can be written to the field CPOL

LBDIER
Possible values of the field LBDIE

LBDIEW
Values that can be written to the field LBDIE

LBDLR
Possible values of the field LBDL

LBDLW
Values that can be written to the field LBDL

LINENR
Possible values of the field LINEN

LINENW
Values that can be written to the field LINEN

STOPR
Possible values of the field STOP

STOPW
Values that can be written to the field STOP