Struct stm32f4xx_hal::timer::Timer

pub struct Timer<TIM> { /* fields omitted */ }

Hardware timers

Implementations

impl Timer<SYST>

pub fn syst<T>(syst: SYST, timeout: T, clocks: Clocks) -> Self where
T: Into<Hertz>,

Configures the SYST clock as a periodic count down timer

pub fn listen(&mut self, event: Event)

Starts listening for an event

pub fn unlisten(&mut self, event: Event)

Stops listening for an event

impl Timer<TIM1>

pub fn tim1<T>(tim: TIM1, timeout: T, clocks: Clocks) -> Self where
T: Into<Hertz>,

Configures a TIM peripheral as a periodic count down timer

pub fn listen(&mut self, event: Event)

Starts listening for an event

Note, you will also have to enable the TIM2 interrupt in the NVIC to start receiving events.

pub fn clear_interrupt(&mut self, event: Event)

Clears interrupt associated with event.

If the interrupt is not cleared, it will immediately retrigger after the ISR has finished.

pub fn unlisten(&mut self, event: Event)

Stops listening for an event

pub fn release(self) -> TIM1

Releases the TIM peripheral

impl Timer<TIM5>

pub fn tim5<T>(tim: TIM5, timeout: T, clocks: Clocks) -> Self where
T: Into<Hertz>,

Configures a TIM peripheral as a periodic count down timer

pub fn listen(&mut self, event: Event)

Starts listening for an event

Note, you will also have to enable the TIM2 interrupt in the NVIC to start receiving events.

pub fn clear_interrupt(&mut self, event: Event)

Clears interrupt associated with event.

If the interrupt is not cleared, it will immediately retrigger after the ISR has finished.

pub fn unlisten(&mut self, event: Event)

Stops listening for an event

pub fn release(self) -> TIM5

Releases the TIM peripheral

impl Timer<TIM9>

pub fn tim9<T>(tim: TIM9, timeout: T, clocks: Clocks) -> Self where
T: Into<Hertz>,

Configures a TIM peripheral as a periodic count down timer

pub fn listen(&mut self, event: Event)

Starts listening for an event

Note, you will also have to enable the TIM2 interrupt in the NVIC to start receiving events.

pub fn clear_interrupt(&mut self, event: Event)

Clears interrupt associated with event.

If the interrupt is not cleared, it will immediately retrigger after the ISR has finished.

pub fn unlisten(&mut self, event: Event)

Stops listening for an event

pub fn release(self) -> TIM9

Releases the TIM peripheral

impl Timer<TIM11>

pub fn tim11<T>(tim: TIM11, timeout: T, clocks: Clocks) -> Self where
T: Into<Hertz>,

Configures a TIM peripheral as a periodic count down timer

pub fn listen(&mut self, event: Event)

Starts listening for an event

Note, you will also have to enable the TIM2 interrupt in the NVIC to start receiving events.

pub fn clear_interrupt(&mut self, event: Event)

Clears interrupt associated with event.

If the interrupt is not cleared, it will immediately retrigger after the ISR has finished.

pub fn unlisten(&mut self, event: Event)

Stops listening for an event

pub fn release(self) -> TIM11

Releases the TIM peripheral

impl Timer<TIM2>

pub fn tim2<T>(tim: TIM2, timeout: T, clocks: Clocks) -> Self where
T: Into<Hertz>,

Configures a TIM peripheral as a periodic count down timer

pub fn listen(&mut self, event: Event)

Starts listening for an event

Note, you will also have to enable the TIM2 interrupt in the NVIC to start receiving events.

pub fn clear_interrupt(&mut self, event: Event)

Clears interrupt associated with event.

If the interrupt is not cleared, it will immediately retrigger after the ISR has finished.

pub fn unlisten(&mut self, event: Event)

Stops listening for an event

pub fn release(self) -> TIM2

Releases the TIM peripheral

impl Timer<TIM3>

pub fn tim3<T>(tim: TIM3, timeout: T, clocks: Clocks) -> Self where
T: Into<Hertz>,

Configures a TIM peripheral as a periodic count down timer

pub fn listen(&mut self, event: Event)

Starts listening for an event

Note, you will also have to enable the TIM2 interrupt in the NVIC to start receiving events.

pub fn clear_interrupt(&mut self, event: Event)

Clears interrupt associated with event.

If the interrupt is not cleared, it will immediately retrigger after the ISR has finished.

pub fn unlisten(&mut self, event: Event)

Stops listening for an event

pub fn release(self) -> TIM3

Releases the TIM peripheral

impl Timer<TIM4>

pub fn tim4<T>(tim: TIM4, timeout: T, clocks: Clocks) -> Self where
T: Into<Hertz>,

Configures a TIM peripheral as a periodic count down timer

pub fn listen(&mut self, event: Event)

Starts listening for an event

Note, you will also have to enable the TIM2 interrupt in the NVIC to start receiving events.

pub fn clear_interrupt(&mut self, event: Event)

Clears interrupt associated with event.

If the interrupt is not cleared, it will immediately retrigger after the ISR has finished.

pub fn unlisten(&mut self, event: Event)

Stops listening for an event

pub fn release(self) -> TIM4

Releases the TIM peripheral

impl Timer<TIM10>

pub fn tim10<T>(tim: TIM10, timeout: T, clocks: Clocks) -> Self where
T: Into<Hertz>,

Configures a TIM peripheral as a periodic count down timer

pub fn listen(&mut self, event: Event)

Starts listening for an event

Note, you will also have to enable the TIM2 interrupt in the NVIC to start receiving events.

pub fn clear_interrupt(&mut self, event: Event)

Clears interrupt associated with event.

If the interrupt is not cleared, it will immediately retrigger after the ISR has finished.

pub fn unlisten(&mut self, event: Event)

Stops listening for an event

pub fn release(self) -> TIM10

Releases the TIM peripheral

impl Timer<TIM6>

pub fn tim6<T>(tim: TIM6, timeout: T, clocks: Clocks) -> Self where
T: Into<Hertz>,

Configures a TIM peripheral as a periodic count down timer

pub fn listen(&mut self, event: Event)

Starts listening for an event

Note, you will also have to enable the TIM2 interrupt in the NVIC to start receiving events.

pub fn clear_interrupt(&mut self, event: Event)

Clears interrupt associated with event.

If the interrupt is not cleared, it will immediately retrigger after the ISR has finished.

pub fn unlisten(&mut self, event: Event)

Stops listening for an event

pub fn release(self) -> TIM6

Releases the TIM peripheral

impl Timer<TIM7>

pub fn tim7<T>(tim: TIM7, timeout: T, clocks: Clocks) -> Self where
T: Into<Hertz>,

Configures a TIM peripheral as a periodic count down timer

pub fn listen(&mut self, event: Event)

Starts listening for an event

Note, you will also have to enable the TIM2 interrupt in the NVIC to start receiving events.

pub fn clear_interrupt(&mut self, event: Event)

Clears interrupt associated with event.

If the interrupt is not cleared, it will immediately retrigger after the ISR has finished.

pub fn unlisten(&mut self, event: Event)

Stops listening for an event

pub fn release(self) -> TIM7

Releases the TIM peripheral

impl Timer<TIM8>

pub fn tim8<T>(tim: TIM8, timeout: T, clocks: Clocks) -> Self where
T: Into<Hertz>,

Configures a TIM peripheral as a periodic count down timer

pub fn listen(&mut self, event: Event)

Starts listening for an event

Note, you will also have to enable the TIM2 interrupt in the NVIC to start receiving events.

pub fn clear_interrupt(&mut self, event: Event)

Clears interrupt associated with event.

If the interrupt is not cleared, it will immediately retrigger after the ISR has finished.

pub fn unlisten(&mut self, event: Event)

Stops listening for an event

pub fn release(self) -> TIM8

Releases the TIM peripheral

impl Timer<TIM12>

pub fn tim12<T>(tim: TIM12, timeout: T, clocks: Clocks) -> Self where
T: Into<Hertz>,

Configures a TIM peripheral as a periodic count down timer

pub fn listen(&mut self, event: Event)

Starts listening for an event

Note, you will also have to enable the TIM2 interrupt in the NVIC to start receiving events.

pub fn clear_interrupt(&mut self, event: Event)

Clears interrupt associated with event.

If the interrupt is not cleared, it will immediately retrigger after the ISR has finished.

pub fn unlisten(&mut self, event: Event)

Stops listening for an event

pub fn release(self) -> TIM12

Releases the TIM peripheral

impl Timer<TIM13>

pub fn tim13<T>(tim: TIM13, timeout: T, clocks: Clocks) -> Self where
T: Into<Hertz>,

Configures a TIM peripheral as a periodic count down timer

pub fn listen(&mut self, event: Event)

Starts listening for an event

Note, you will also have to enable the TIM2 interrupt in the NVIC to start receiving events.

pub fn clear_interrupt(&mut self, event: Event)

Clears interrupt associated with event.

If the interrupt is not cleared, it will immediately retrigger after the ISR has finished.

pub fn unlisten(&mut self, event: Event)

Stops listening for an event

pub fn release(self) -> TIM13

Releases the TIM peripheral

impl Timer<TIM14>

pub fn tim14<T>(tim: TIM14, timeout: T, clocks: Clocks) -> Self where
T: Into<Hertz>,

Configures a TIM peripheral as a periodic count down timer

pub fn listen(&mut self, event: Event)

Starts listening for an event

Note, you will also have to enable the TIM2 interrupt in the NVIC to start receiving events.

pub fn clear_interrupt(&mut self, event: Event)

Clears interrupt associated with event.

If the interrupt is not cleared, it will immediately retrigger after the ISR has finished.

pub fn unlisten(&mut self, event: Event)

Stops listening for an event

pub fn release(self) -> TIM14

Releases the TIM peripheral