Module stm32f4xx_hal::dma::traits

Structs

ARR

Wrapper type that indicates which register of the contained timer to use for DMA.

CCR1

Wrapper type that indicates which register of the contained timer to use for DMA.

CCR2

Wrapper type that indicates which register of the contained timer to use for DMA.

CCR3

Wrapper type that indicates which register of the contained timer to use for DMA.

CCR4

Wrapper type that indicates which register of the contained timer to use for DMA.

DMAR

Wrapper type that indicates which register of the contained timer to use for DMA.

Traits

Channel

A channel that can be configured on a DMA stream.

DMASet

Trait to mark a set of Stream, Channel and Direction for a Peripheral as correct together.

Direction

DMA direction.

Instance

Trait that represents an instance of a DMA peripheral.

PeriAddress

Get an address and memory size the DMA can use.

SafePeripheralRead

Marker trait for structs which can be safely accessed with shared reference

Stream

Trait for DMA streams types.

StreamISR

Trait for DMA stream interrupt handling.

Type Definitions

DMARegisterBlock

Type alias to a DMA RegisterBlock.