Trait stm32f4xx_hal::dma::traits::PeriAddress

pub unsafe trait PeriAddress {
    type MemSize;
    fn address(&self) -> u32;
}

Get an address and memory size the DMA can use.

Safety

Both the memory size and the address must be correct for the specific peripheral and for the DMA.

Associated Types

type MemSize

Memory size of the peripheral.

Required methods

fn address(&self) -> u32

Returns the address to be used by the DMA stream.