Module stm32f4xx_hal::spi

Structs

DmaBuilder

Miso

Mode

SPI mode

Mosi

Nss

Rx

Sck

Spi

TransferModeBidi

BIDI mode - use TX pin as RX then spi receive data

TransferModeNormal

Normal mode - RX and TX pins are independent

Tx

Enums

Error

SPI error

Event

Interrupt events

Phase

Clock phase

Polarity

Clock polarity

Traits

Instance

Pins

SpiExt

Type Definitions

NoMiso

A filler type for when the Miso pin is unnecessary

NoMosi

A filler type for when the Mosi pin is unnecessary

NoSck

A filler type for when the SCK pin is unnecessary

Spi1

Spi2

Spi3

Spi4

Spi5