Struct stm32f4xx_hal::spi::Spi

pub struct Spi<SPI, PINS, TRANSFER_MODE> { /* fields omitted */ }

Implementations

impl<SPI, PINS> Spi<SPI, PINS, TransferModeNormal> where
SPI: Instance,
PINS: Pins<SPI>,

pub fn new(
spi: SPI,
pins: PINS,
mode: impl Into<Mode>,
freq: Hertz,
clocks: &Clocks
) -> Self

pub fn to_bidi_transfer_mode(self) -> Spi<SPI, PINS, TransferModeBidi>

impl<SPI, PINS> Spi<SPI, PINS, TransferModeBidi> where
SPI: Instance,
PINS: Pins<SPI>,

pub fn new_bidi(
spi: SPI,
pins: PINS,
mode: impl Into<Mode>,
freq: Hertz,
clocks: &Clocks
) -> Self

pub fn to_normal_transfer_mode(self) -> Spi<SPI, PINS, TransferModeNormal>

impl<SPI, PINS, TRANSFER_MODE> Spi<SPI, PINS, TRANSFER_MODE> where
SPI: Instance,
PINS: Pins<SPI>,

pub fn release(self) -> (SPI, PINS)

impl<SPI, PINS> Spi<SPI, PINS, TransferModeNormal> where
SPI: Instance,

pub fn init(self) -> Self

impl<SPI, PINS> Spi<SPI, PINS, TransferModeBidi> where
SPI: Instance,

pub fn init(self) -> Self

impl<SPI, PINS, TRANSFER_MODE> Spi<SPI, PINS, TRANSFER_MODE> where
SPI: Instance,

pub fn enable(&mut self, enable: bool)

Enable/disable spi

pub fn listen(&mut self, event: Event)

Enable interrupts for the given event:

Received data ready to be read (RXNE)

Transmit data register empty (TXE)

Transfer error

pub fn unlisten(&mut self, event: Event)

Disable interrupts for the given event:

Received data ready to be read (RXNE)

Transmit data register empty (TXE)

Transfer error

pub fn is_txe(&self) -> bool

Return true if the TXE flag is set, i.e. new data to transmit can be written to the SPI.

pub fn is_rxne(&self) -> bool

Return true if the RXNE flag is set, i.e. new data has been received and can be read from the SPI.

pub fn is_modf(&self) -> bool

Return true if the MODF flag is set, i.e. the SPI has experienced a Master Mode Fault. (see chapter 28.3.10 of the STM32F4 Reference Manual)

pub fn is_ovr(&self) -> bool

Return true if the OVR flag is set, i.e. new data has been received while the receive data register was already filled.

pub fn use_dma(self) -> DmaBuilder<SPI>