Struct stm32f4xx_hal::spi::Spi

pub struct Spi<SPI, PINS> { /* fields omitted */ }

Methods

impl<PINS> Spi<SPI1, PINS>

pub fn spi1(
spi: SPI1,
pins: PINS,
mode: Mode,
freq: Hertz,
clocks: Clocks
) -> Self where
PINS: Pins<SPI1>,

pub fn listen(&mut self, event: Event)

Enable interrupts for the given event:

Received data ready to be read (RXNE)

Transmit data register empty (TXE)

Transfer error

pub fn unlisten(&mut self, event: Event)

Disable interrupts for the given event:

Received data ready to be read (RXNE)

Transmit data register empty (TXE)

Transfer error

pub fn is_txe(&self) -> bool

Return true if the TXE flag is set, i.e. new data to transmit can be written to the SPI.

pub fn is_rxne(&self) -> bool

Return true if the RXNE flag is set, i.e. new data has been received and can be read from the SPI.

pub fn is_modf(&self) -> bool

Return true if the MODF flag is set, i.e. the SPI has experienced a Master Mode Fault. (see chapter 28.3.10 of the STM32F4 Reference Manual)

pub fn is_ovr(&self) -> bool

Return true if the OVR flag is set, i.e. new data has been received while the receive data register was already filled.

pub fn free(self) -> (SPI1, PINS)

impl<PINS> Spi<SPI2, PINS>

pub fn spi2(
spi: SPI2,
pins: PINS,
mode: Mode,
freq: Hertz,
clocks: Clocks
) -> Self where
PINS: Pins<SPI2>,

pub fn listen(&mut self, event: Event)

Enable interrupts for the given event:

Received data ready to be read (RXNE)

Transmit data register empty (TXE)

Transfer error

pub fn unlisten(&mut self, event: Event)

Disable interrupts for the given event:

Received data ready to be read (RXNE)

Transmit data register empty (TXE)

Transfer error

pub fn is_txe(&self) -> bool

Return true if the TXE flag is set, i.e. new data to transmit can be written to the SPI.

pub fn is_rxne(&self) -> bool

Return true if the RXNE flag is set, i.e. new data has been received and can be read from the SPI.

pub fn is_modf(&self) -> bool

Return true if the MODF flag is set, i.e. the SPI has experienced a Master Mode Fault. (see chapter 28.3.10 of the STM32F4 Reference Manual)

pub fn is_ovr(&self) -> bool

Return true if the OVR flag is set, i.e. new data has been received while the receive data register was already filled.

pub fn free(self) -> (SPI2, PINS)

impl<PINS> Spi<SPI3, PINS>

pub fn spi3(
spi: SPI3,
pins: PINS,
mode: Mode,
freq: Hertz,
clocks: Clocks
) -> Self where
PINS: Pins<SPI3>,

pub fn listen(&mut self, event: Event)

Enable interrupts for the given event:

Received data ready to be read (RXNE)

Transmit data register empty (TXE)

Transfer error

pub fn unlisten(&mut self, event: Event)

Disable interrupts for the given event:

Received data ready to be read (RXNE)

Transmit data register empty (TXE)

Transfer error

pub fn is_txe(&self) -> bool

Return true if the TXE flag is set, i.e. new data to transmit can be written to the SPI.

pub fn is_rxne(&self) -> bool

Return true if the RXNE flag is set, i.e. new data has been received and can be read from the SPI.

pub fn is_modf(&self) -> bool

Return true if the MODF flag is set, i.e. the SPI has experienced a Master Mode Fault. (see chapter 28.3.10 of the STM32F4 Reference Manual)

pub fn is_ovr(&self) -> bool

Return true if the OVR flag is set, i.e. new data has been received while the receive data register was already filled.

pub fn free(self) -> (SPI3, PINS)

impl<PINS> Spi<SPI4, PINS>

pub fn spi4(
spi: SPI4,
pins: PINS,
mode: Mode,
freq: Hertz,
clocks: Clocks
) -> Self where
PINS: Pins<SPI4>,

pub fn listen(&mut self, event: Event)

Enable interrupts for the given event:

Received data ready to be read (RXNE)

Transmit data register empty (TXE)

Transfer error

pub fn unlisten(&mut self, event: Event)

Disable interrupts for the given event:

Received data ready to be read (RXNE)

Transmit data register empty (TXE)

Transfer error

pub fn is_txe(&self) -> bool

Return true if the TXE flag is set, i.e. new data to transmit can be written to the SPI.

pub fn is_rxne(&self) -> bool

Return true if the RXNE flag is set, i.e. new data has been received and can be read from the SPI.

pub fn is_modf(&self) -> bool

Return true if the MODF flag is set, i.e. the SPI has experienced a Master Mode Fault. (see chapter 28.3.10 of the STM32F4 Reference Manual)

pub fn is_ovr(&self) -> bool

Return true if the OVR flag is set, i.e. new data has been received while the receive data register was already filled.

pub fn free(self) -> (SPI4, PINS)

impl<PINS> Spi<SPI5, PINS>

pub fn spi5(
spi: SPI5,
pins: PINS,
mode: Mode,
freq: Hertz,
clocks: Clocks
) -> Self where
PINS: Pins<SPI5>,

pub fn listen(&mut self, event: Event)

Enable interrupts for the given event:

Received data ready to be read (RXNE)

Transmit data register empty (TXE)

Transfer error

pub fn unlisten(&mut self, event: Event)

Disable interrupts for the given event:

Received data ready to be read (RXNE)

Transmit data register empty (TXE)

Transfer error

pub fn is_txe(&self) -> bool

Return true if the TXE flag is set, i.e. new data to transmit can be written to the SPI.

pub fn is_rxne(&self) -> bool

Return true if the RXNE flag is set, i.e. new data has been received and can be read from the SPI.

pub fn is_modf(&self) -> bool

Return true if the MODF flag is set, i.e. the SPI has experienced a Master Mode Fault. (see chapter 28.3.10 of the STM32F4 Reference Manual)

pub fn is_ovr(&self) -> bool

Return true if the OVR flag is set, i.e. new data has been received while the receive data register was already filled.

pub fn free(self) -> (SPI5, PINS)